Refactored code into hopefully clearer parts
